refactor swift compiler for better inference
